requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
cffi==1.12.2
Click==7.0
colorama==0.4.0
coverage==4.5.2
cryptography==2.6.1
flake8==3.6.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.13.1
gunicorn==19.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.6.4
pylint==2.1.1
pytest==4.0.0
six==1.12.0
typed-ast==1.1.0
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
